backend/requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
image==1.5.33
pillow==10.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
python-decouple==3.8
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
redis==5.0.1
django-redis==5.4.0
django-storages==1.14.2
boto3==1.34.0
sentry-sdk==2.3.1
celery==5.3.6
django-celery-beat==2.6.0
openai>=1.30.0
